# Dbscan nearest neighbour integration