--- a/algorithms/linfa-clustering/src/appx_dbscan/clustering/mod.rs
+++ b/algorithms/linfa-clustering/src/appx_dbscan/clustering/mod.rs
@@ -39,11 +39,11 @@ impl AppxDbscanLabeler {
     /// ```rust
     ///
     /// use ndarray::{array, Axis};
-    /// use linfa_clustering::{AppxDbscanLabeler, AppxDbscanHyperParams};
+    /// use linfa_clustering::{AppxDbscanLabeler, AppxDbscan};
     ///
     /// // Let's define some observations and set the desired params
     /// let observations = array![[0.,0.], [1., 0.], [0., 1.]];
-    /// let params = AppxDbscanHyperParams::new(2).build();
+    /// let params = AppxDbscan::params(2);
     /// // Now we build the labels for each observation using the Labeler struct
     /// let labeler = AppxDbscanLabeler::new(&observations.view(),&params);
     /// // Here we can access the labels for each point `observations`
@@ -54,6 +54,10 @@ impl AppxDbscanLabeler {
     ///
     pub fn labels(&self) -> &Array1<Option<usize>> {
         &self.labels
+    }
+
+    pub(crate) fn into_labels(self) -> Array1<Option<usize>> {
+        self.labels
     }
 
     fn label<F: Float>(