--- a/algorithms/linfa-clustering/src/appx_dbscan/tests.rs
+++ b/algorithms/linfa-clustering/src/appx_dbscan/tests.rs
@@ -1,5 +1,5 @@
 use crate::{generate_blobs, AppxDbscan, AppxDbscanHyperParams, Dbscan};
-use linfa::traits::Transformer;
+use linfa::traits::Predict;
 use ndarray::{arr2, s, Array1, Array2};
 use ndarray_rand::rand::SeedableRng;
 use ndarray_rand::rand_distr::Uniform;
@@ -19,10 +19,10 @@ fn appx_dbscan_test_100() {
     let appx_res = AppxDbscan::params(min_points)
         .tolerance(tolerance)
         .slack(1e-4)
-        .transform(&dataset);
+        .predict(&dataset);
     let ex_res = Dbscan::params(min_points)
         .tolerance(tolerance)
-        .transform(&dataset);
+        .predict(&dataset);
 
     // The order of the labels of the clusters in the two algorithms may not be the same
     // but it does not affect the result. We have to create a mapping from the exact labels
@@ -67,10 +67,10 @@ fn appx_dbscan_test_250() {
     let appx_res = AppxDbscan::params(min_points)
         .tolerance(tolerance)
         .slack(1e-4)
-        .transform(&dataset);
+        .predict(&dataset);
     let ex_res = Dbscan::params(min_points)
         .tolerance(tolerance)
-        .transform(&dataset);
+        .predict(&dataset);
 
     // The order of the labels of the clusters in the two algorithms may not be the same
     // but it does not affect the result. We have to create a mapping from the exact labels
@@ -112,10 +112,10 @@ fn appx_dbscan_test_500() {
     let appx_res = AppxDbscan::params(min_points)
         .tolerance(tolerance)
         .slack(1e-4)
-        .transform(&dataset);
+        .predict(&dataset);
     let ex_res = Dbscan::params(min_points)
         .tolerance(tolerance)
-        .transform(&dataset);
+        .predict(&dataset);
 
     // The order of the labels of the clusters in the two algorithms may not be the same
     // but it does not affect the result. We have to create a mapping from the exact labels
@@ -163,7 +163,7 @@ fn test_border() {
     let labels = AppxDbscan::params(5)
         .tolerance(1.1)
         .slack(1e-5)
-        .transform(&data);
+        .predict(&data);
 
     assert_eq!(labels[0], None);
     for id in labels.slice(s![1..]).iter() {
@@ -213,7 +213,7 @@ fn test_outliers() {
     let labels = AppxDbscan::params(2)
         .tolerance(1.0)
         .slack(1e-4)
-        .transform(&data);
+        .predict(&data);
     // we should find that the first 50 points are all in the same cluster (cluster 0)
     // and that the other points are so far away from one another that they are all noise points
     for i in 0..50 {
@@ -246,7 +246,7 @@ fn nested_clusters() {
     let labels = AppxDbscan::params(2)
         .tolerance(1.0)
         .slack(1e-4)
-        .transform(&data);
+        .predict(&data);
 
     assert!(labels.slice(s![..40]).iter().all(|x| x == &Some(0)));
     assert!(labels.slice(s![40..]).iter().all(|x| x == &Some(1)));
@@ -276,44 +276,29 @@ fn test_exp(){
 #[test]
 #[should_panic]
 fn tolerance_cannot_be_zero() {
-    AppxDbscanHyperParams::new(2)
-        .tolerance(0.0)
-        .slack(0.1)
-        .build();
+    AppxDbscanHyperParams::new(2).tolerance(0.0).slack(0.1);
 }
 
 #[test]
 #[should_panic]
 fn slack_cannot_be_zero() {
-    AppxDbscanHyperParams::new(2)
-        .tolerance(0.1)
-        .slack(0.0)
-        .build();
+    AppxDbscanHyperParams::new(2).tolerance(0.1).slack(0.0);
 }
 
 #[test]
 #[should_panic]
 fn min_points_at_least_2() {
-    AppxDbscanHyperParams::new(1)
-        .tolerance(0.1)
-        .slack(0.1)
-        .build();
+    AppxDbscanHyperParams::new(1).tolerance(0.1).slack(0.1);
 }
 
 #[test]
 #[should_panic]
 fn tolerance_should_be_positive() {
-    AppxDbscanHyperParams::new(2)
-        .tolerance(-1.0)
-        .slack(0.1)
-        .build();
+    AppxDbscanHyperParams::new(2).tolerance(-1.0).slack(0.1);
 }
 
 #[test]
 #[should_panic]
 fn slack_should_be_positive() {
-    AppxDbscanHyperParams::new(2)
-        .tolerance(0.1)
-        .slack(-1.0)
-        .build();
+    AppxDbscanHyperParams::new(2).tolerance(0.1).slack(-1.0);
 }