Allow DBSCAN to use nearest neighbour algorithms for range query. The only major breaking change to DBSCAN is the removal of the `Transformer` impl for `DbscanHyperparamBuilder`. The impl didn't really make sense on a builder struct and didn't play well with NearestNeighbour structs. The `Clone` and `PartialEq` derives on the hyperparameters were also removed since they didn't work with NearestNeighbour and are not needed to run the algorithm. Breaking changes are also required for `linfa-nn` outlined [here](https://github.com/rust-ml/linfa/issues/119#issuecomment-841839071).