@YuhanLiin 2021-05-19
Turns out serde derives don't work really well with trait objects, and `Deserialize` doesn't work at all even with macros like `typetag`. Might want to use an enum instead of trait object to represent `NearestNeighbour` types.