@YuhanLiin 2021-05-20
Changed NearestNeighbour trait objects to an enum and everything works fine. Will run benchmarks soon