@YuhanLiin 2021-05-22
Compared the benchmarks and the differences were insane.
Master benchmarks:
```
dbscan/dbscan/10        time:   [28.040 us 28.088 us 28.137 us]                              
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
dbscan/dbscan/100       time:   [2.6009 ms 2.6080 ms 2.6169 ms]                               
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
Benchmarking dbscan/dbscan/1000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 27.4s, or reduce sample count to 10.
dbscan/dbscan/1000      time:   [269.90 ms 270.38 ms 271.02 ms]                               
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
```
Note that I didn't include the results for `dbscan/10000` since that was taking way too long.

Benchmark with KD Tree:
```
dbscan/dbscan/10        time:   [26.876 us 27.045 us 27.240 us]                              
                        change: [-4.5658% -2.9422% -1.5433%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe
dbscan/dbscan/100       time:   [653.07 us 660.53 us 668.45 us]                              
                        change: [-74.642% -74.385% -74.103%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
dbscan/dbscan/1000      time:   [14.886 ms 14.977 ms 15.082 ms]                               
                        change: [-94.497% -94.461% -94.422%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
Benchmarking dbscan/dbscan/10000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 48.7s, or reduce sample count to 10.
dbscan/dbscan/10000     time:   [465.09 ms 466.72 ms 468.52 ms]                                
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
```
We might want to make `balltree` the default instead of `kdtree` because `kdtree` panics on non-contiguous data points, but even then the speed difference is still remarkable.