@YuhanLiin 2021-06-09
How should I handle the errors thrown by the NN operations?