@bytesnake 2021-06-10
> How should I handle the errors thrown by the NN operations?

DBSCAN should have its own error variant which can wrap the NN errors